int -> long dynamicStash

+++ dynStashplus/TwoChoiceClient.h
@@ -24,25 +24,25 @@ private:
 
 public:
     virtual ~TwoChoiceClient();
-    TwoChoiceClient(int maxUpdate, bool inMemory, bool overwrite, bool profile);
-    int totalCommunication = 0;
-    vector<int> numberOfBins;
+    TwoChoiceClient(long maxUpdate, bool inMemory, bool overwrite, bool profile);
+    long totalCommunication = 0;
+    vector<long> numberOfBins;
     prf_type nullKey;
-    vector<int> sizeOfEachBin;
-    map<int,int> position;
+    vector<long> sizeOfEachBin;
+    map<long,long> position;
     vector<bool> exist;
     vector<bool> existone;
-    void destry(int index);
-    void setup(int index, map<string, vector<prf_type> >pairs, unsigned char* key);
-    vector<prf_type> search(int index, string keyword, unsigned char* key);
-    vector<prf_type> newsearch(int index, string keyword, unsigned char* key);
-    vector<prf_type> getAllData(int index, unsigned char* key);
-    void truncateToMpl(int pss, int mpl, int index, string keyword, vector<prf_type> fileids, unsigned char* key);
-    void writeToCuckooHT(int index, int mpl, string keyword, vector<prf_type> fileids, unsigned char* key);
-    void writeToCuckooStash(vector<prf_type> fileids,int cnt,int index, int tableNum, unsigned char* key);
-void place(string keyw,vector<prf_type> fileids,int cuckooID,int cnt,int index, int tableNum, unsigned char* key);
-void writeToStash(int pss,int mpl,vector<prf_type> fileids,unsigned char* key,vector<prf_type> &stashCiphers);
-int maxPossibleLen(int index);
+    void destry(long index);
+    void setup(long index, map<string, vector<prf_type> >pairs, unsigned char* key);
+    vector<prf_type> search(long index, string keyword, unsigned char* key);
+    vector<prf_type> newsearch(long index, string keyword, unsigned char* key);
+    vector<prf_type> getAllData(long index, unsigned char* key);
+    void truncateToMpl(long pss, long mpl, long index, string keyword, vector<prf_type> fileids, unsigned char* key);
+    void writeToCuckooHT(long index, long mpl, string keyword, vector<prf_type> fileids, unsigned char* key);
+    void writeToCuckooStash(vector<prf_type> fileids,long cnt,long index, long tableNum, unsigned char* key);
+void place(string keyw,vector<prf_type> fileids,long cuckooID,long cnt,long index, long tableNum, unsigned char* key);
+void writeToStash(long pss,long mpl,vector<prf_type> fileids,unsigned char* key,vector<prf_type> &stashCiphers);
+long maxPossibleLen(long index);
 void printStashSizes();
 };
